项目提交

+++ onlineQuestionnaireSurvey/src/main/java/com/sicnu/baby/web/QuestionnaireController.java
@@ -2,7 +2,6 @@ package com.sicnu.baby.web;
 
 
 import javax.annotation.Resource;
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import org.json.JSONObject;
@@ -48,14 +47,14 @@ public class QuestionnaireController {
 	 * @return
 	 */
 	@RequestMapping(value="/put/questionnaire/private",
-			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"},
+			method=RequestMethod.POST,
+			produces={"application/json;charset=UTF-8"},
 			headers={"content-type=application/json;charset=UTF-8"})
 	public @ResponseBody QuestionnaireResult<Url> addQuestionnaire(@RequestBody JSONObject json,HttpSession session){
 		try{
 		   User user = (User)session.getAttribute("user");
 		   if(questionnaireService.addQuestionnaire(json,user.getUsername())){
-			   return new QuestionnaireResult<>(true,new Url(""),1);
+			   return new QuestionnaireResult<>(true,new Url("questionnaire.html"),1);
 		   }
 		   return new QuestionnaireResult<>(false, "添加失败");
 		}catch (QuesSurveyDataException e) {
@@ -73,50 +72,52 @@ public class QuestionnaireController {
 	 * @return
 	 */
 	@RequestMapping(value="/delete/{quesId}/questionnaire/private",
-			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
+			method=RequestMethod.POST,
+			produces={"application/json;charset=UTF-8"})
 	public @ResponseBody QuestionnaireResult<String> deleteQuestionnaireByUser(@PathVariable("quesId") int quesId,HttpSession session){
 		User user = (User)session.getAttribute("user");
-		if(questionnaireService.deleteById(quesId,user.getUsername())){
-			return new QuestionnaireResult<>(true, null);
+		if(user.getUserType()==0){
+			if(questionnaireService.deleteById(quesId,user.getUsername())){
+				return new QuestionnaireResult<>(true, null);
+			}else{
+				return new QuestionnaireResult<>(false, "删除失败");
+			}			
 		}else{
-			return new QuestionnaireResult<>(true, "删除失败");
+			if(questionnaireService.deleteById(quesId,null)){
+				return new QuestionnaireResult<>(true, null);
+			}else{
+				return new QuestionnaireResult<>(false, "删除失败");
+			}
 		}
 	}
 	
 	/**
-	 * 管理员删除问卷
-	 * @param quesId
-	 * @return
-	 */
-	@RequestMapping(value="/delete/{quesId}/questionnaire/admin/private",
-			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
-	public @ResponseBody QuestionnaireResult<String> deleteQuestionnaireById(@PathVariable("quesId") int quesId){
-		if(questionnaireService.deleteById(quesId,null)){
-			return new QuestionnaireResult<>(true, null);
-		}else{
-			return new QuestionnaireResult<>(true, "删除失败");
-		}
-	}
-	/**
-	 * 用于问卷发起者查看问卷
+	 * 用于查看问卷
 	 * @param session
 	 * @return
 	 */
-	@RequestMapping(value="/get/user/questionnaire/private",
+	@RequestMapping(value="/get/page/questionnaire/private",
 			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
+			produces={"application/json;charset=UTF-8"})
 	public @ResponseBody QuestionnaireResult<PageAndQuestionnaire> getQuestionnaire(HttpSession session){
 		User user = (User)session.getAttribute("user");
 		Page page = new Page();
-		page.setPageNumber(20);
+		page.setPageNumber(8);
 		page.setCurrentPage(1);
-		PageAndQuestionnaire questionnaires = questionnaireService.getQuestionByUsername(user.getUsername(),page);
-		if(questionnaires!=null){
-			return new QuestionnaireResult<>(true, questionnaires, 1);
+		if(user.getUserType()==0){
+			PageAndQuestionnaire questionnaires = questionnaireService.getQuestionByUsername(user.getUsername(),page);
+			if(questionnaires!=null){
+				return new QuestionnaireResult<>(true, questionnaires, 1);
+			}else{
+				return new QuestionnaireResult<>(false, "未查询到任何数据");
+			}			
 		}else{
-			return new QuestionnaireResult<>(false, "未查询到任何数据");
+			PageAndQuestionnaire pQuestionnaire = questionnaireService.getAllQuestion(page);
+			if(pQuestionnaire!=null){
+				return new QuestionnaireResult<PageAndQuestionnaire>(true, pQuestionnaire, 1);			
+			}else{
+				return new QuestionnaireResult<>(false, "未查询到任何数据");
+			}
 		}
 	}
 	
@@ -127,32 +128,12 @@ public class QuestionnaireController {
 	 */
 	@RequestMapping(value="/get/detail/{quesId}/questionnaire",
 			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
+			produces={"application/json;charset=UTF-8"})
 	public @ResponseBody QuestionnaireResult<Questionnaire> getQuestionnaireDetail(@PathVariable("quesId")  int quesId){
 		Questionnaire questionnaire = questionnaireService.getQuestionnaireById(quesId);
 		if(questionnaire!=null){
 			questionnaire.setQesResult(null);
 			return new QuestionnaireResult<>(true, questionnaire,1);
-		}else{
-			return new QuestionnaireResult<>(false, "未查询到任何数据");
-		}
-	}
-	/**
-	 * 初始化分页信息
-	 * @param request
-	 * @return
-	 */
-	@RequestMapping(value="/get/page/questionnaire/admin/private",
-			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
-	public @ResponseBody QuestionnaireResult<PageAndQuestionnaire> getAllQuestionnaire(HttpServletRequest request){
-		Page page = new Page();
-		page.setPageNumber(1);
-		page.setCurrentPage(1);
-		PageAndQuestionnaire pQuestionnaire = questionnaireService.getAllQuestion(page);
-		if(pQuestionnaire!=null){
-			request.setAttribute("page", page);
-			return new QuestionnaireResult<PageAndQuestionnaire>(true, pQuestionnaire, 1);			
 		}else{
 			return new QuestionnaireResult<>(false, "未查询到任何数据");
 		}
@@ -163,17 +144,27 @@ public class QuestionnaireController {
 	 * @param request
 	 * @return
 	 */
-	@RequestMapping(value="/get/page/{index}/questionnaire/admin/private",
+	@RequestMapping(value="/get/page/{index}/questionnaire/private",
 			method=RequestMethod.GET,
-			produces={"application/json;chaeset=UTF-8"})
-	public @ResponseBody QuestionnaireResult<PageAndQuestionnaire> getQuestionnaireByIndex(HttpServletRequest request,@PathVariable("index") int index){
+			produces={"application/json;charset=UTF-8"})
+	public @ResponseBody QuestionnaireResult<PageAndQuestionnaire> getQuestionnaireByIndex(HttpSession session,@PathVariable("index") int index){
+		User user = (User)session.getAttribute("user");
 		Page page = new Page();
 		page.setCurrentPage(index);
-		PageAndQuestionnaire pQuestionnaire = questionnaireService.getAllQuestion(page);
-		if(pQuestionnaire!=null){
-			return new QuestionnaireResult<PageAndQuestionnaire>(true, pQuestionnaire, 1);			
+		if(user.getUserType()==1){
+			PageAndQuestionnaire pQuestionnaire = questionnaireService.getAllQuestion(page);
+			if(pQuestionnaire!=null){
+				return new QuestionnaireResult<PageAndQuestionnaire>(true, pQuestionnaire, 1);			
+			}else{
+				return new QuestionnaireResult<>(false, "未查询到任何数据");
+			}			
 		}else{
-			return new QuestionnaireResult<>(false, "未查询到任何数据");
+			PageAndQuestionnaire pQuestionnaire = questionnaireService.getQuestionByUsername(user.getUsername(),page);
+			if(pQuestionnaire!=null){
+				return new QuestionnaireResult<PageAndQuestionnaire>(true, pQuestionnaire, 1);			
+			}else{
+				return new QuestionnaireResult<>(false, "未查询到任何数据");
+			}
 		}
 	}
